+++ lib/skylight/util/logging.rb
@@ -5,11 +5,11 @@ module Skylight
     module Logging
       if ENV[TRACE_ENV_KEY]
         def trace(msg, *args)
-          log :DEBUG, msg, *args
+          log :debug, msg, *args
         end
 
         def t
-          log :DEBUG, yield
+          log :debug, yield
         end
       else
         def trace(*)
@@ -20,38 +20,34 @@ module Skylight
       end
 
       def debug(msg, *args)
-        log :DEBUG, msg, *args
+        log :debug, msg, *args
       end
 
       def info(msg, *args)
-        log :INFO, msg, *args
+        log :info, msg, *args
       end
 
       def warn(msg, *args)
-        log :WARN, msg, *args
+        log :warn, msg, *args
       end
 
       def error(msg, *args)
-        log :ERROR, msg, *args
+        log :error, msg, *args
       end
 
       alias fmt sprintf
-
-      MAP = {
-        :DEBUG => Logger::DEBUG,
-        :INFO  => Logger::INFO,
-        :WARN  => Logger::WARN,
-        :ERROR => Logger::ERROR }
 
       def log(level, msg, *args)
         return unless respond_to?(:config)
         return unless c = config
 
         if logger = c.logger
+          return unless logger.respond_to?(level)
+
           if args.length > 0
-            logger.log MAP[level], sprintf("[SKYLIGHT] #{msg}", *args)
+            logger.send level, sprintf("[SKYLIGHT] #{msg}", *args)
           else
-            logger.log MAP[level], "[SKYLIGHT] #{msg}"
+            logger.send level, "[SKYLIGHT] #{msg}"
           end
         end
       rescue Exception => e